Fixed a couple of things with ipynb file
Fixed import.
Added import matplotlib.pyplot as plt.
Fixed folder_name needing to jump one dir back before finding the correct file.
Removed {geom_desc} since it was not needed and caused issues with finding the correct file.
Still an issue with trying to plot outflows_root and outflows_leaf2, need to find out what they should be.
Changed ix as it seems to be connected to the index of which experiment results to look at.