Made changes to generate_simple_arterial_tree.
When a point would collide, it was not added, and this introduced an issue where we would not have enough equations. Now the function does two phases: one where it tries to place as many points as are legal. In the second pass, it checks how many junctions we expect, and if we don't have enough, we try adding until it is sufficient.